api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pdfplumber==0.11.4
PyMuPDF==1.24.9
openpyxl==3.1.5
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
tenacity==8.5.0
orjson==3.10.7
numpy==2.0.2
pandas==2.2.3
google-generativeai==0.8.2
mangum==0.17.0

